Меняет структуру страницы избранное
