+++ src/components/favorites-empty/favorites-empty.tsx
@@ -1,14 +1,18 @@
 function FavoritesEmpty(): JSX.Element {
   return (
-    <section className="favorites favorites--empty">
-      <h1 className="visually-hidden">Favorites (empty)</h1>
-      <div className="favorites__status-wrapper">
-        <b className="favorites__status">Nothing yet saved.</b>
-        <p className="favorites__status-description">
-          Save properties to narrow down search or plan your future trips.
-        </p>
+    <main className="page__main page__main--favorites page__main--favorites-empty">
+      <div className="page__favorites-container container">
+        <section className="favorites favorites--empty">
+          <h1 className="visually-hidden">Favorites (empty)</h1>
+          <div className="favorites__status-wrapper">
+            <b className="favorites__status">Nothing yet saved.</b>
+            <p className="favorites__status-description">
+              Save properties to narrow down search or plan your future trips.
+            </p>
+          </div>
+        </section>
       </div>
-    </section>
+    </main>
   );
 }
